requirements.txt
google-api-python-client~=2.154.0
google-auth~=2.36.0
python-dotenv~=1.0.1
google_auth_oauthlib~=1.2.1
jinja2~=3.1.4